lexer done and organized
