+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: tsoares- <tsoares-@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/18 08:11:33 by tsoares-          #+#    #+#             */
-/*   Updated: 2024/07/10 21:57:42 by jgils            ###   ########.fr       */
+/*   Updated: 2024/07/10 22:36:56 by jgils            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ char	*read_input(char *user_input);
 bool	validate_input(char *input);
 
 
-// PARSER.C -------------------------------------------------------------------
+// TOKENIZER.C ----------------------------------------------------------------
 /*
  * count the amount of tokens in an input
  */
@@ -41,7 +41,24 @@ int	count_tokens(char *input);
 char	**tokenizer(char *user_input);
 
 
-// PARSER_UTILS.C -------------------------------------------------------------
+// TOKENIZER_UTILS.C ----------------------------------------------------------
 char	*strtrim_space(char *s1);
+
+/*
+* ft_substr but starting by the first index of str
+*/
+char	*substr_by_address(char *str, int len);
+
+// PRINT.C --------------------------------------------------------------------
+void	print_split(char **array);
+
+// FREE.C ---------------------------------------------------------------------
+void	free_split(char **array);
+
+// IDENTIFY_CHAR.C ------------------------------------------------------------
+int	is_special(char c);
+int	is_quote(char c);
+int	is_space(char c);
+int	is_basic(char c);
 
 #endif
