+++ srcs/free.c
@@ -1,0 +1,13 @@
+#include "../includes/minishell.h"
+
+void	free_split(char **array)
+{
+	int	i = 0;
+
+	while (array[i] != NULL)
+	{
+		free(array[i]);
+		i++;
+	}
+	free(array);
+}
