+++ srcs/identify_char.c
@@ -1,0 +1,19 @@
+int	is_special(char c)
+{
+	return (c == '|' || c == '<' || c == '>');
+}
+
+int	is_quote(char c)
+{
+	return (c == '\'' || c == '\"');
+}
+
+int	is_space(char c)
+{
+	return ((c >= 9 && c <= 13) || c == 32);
+}
+
+int	is_basic(char c)
+{
+	return (!(is_special(c)) && !(is_quote(c)) && !(is_space(c)));
+}
